Change maintainer to be GitHub.
Now that there is a GitHub repo, it seems sensible to move
maintainership from Simon M to that GitHub repo. Apologies if this
is the wrong thing to do, but it seems that Simon M has not yet
committed to the repo since its creation this spring, leading me
to believe others have taken over.